Move sprite to sprite.py
