+++ arcade/sprite/sprite.py
